refactor: move carina-cli tests from main.rs to tests module
Extract the ~1270 lines of test code from main.rs into a separate
src/tests.rs file. This reduces main.rs from 1452 lines to 183 lines,
making it easier to navigate. The tests module is referenced via
`#[cfg(test)] mod tests;` which loads the external file.

Part of #777
